refactor(tests): move extreme harness library into tests/extreme_harness/ (regular CI gate)
The 22 fast, stdlib-only unit tests for the extreme harness (fault_proxy,
measurement, _storage_discovery) were unmarked but lived under tests-extensive/,
so no CI job ran them. Move the harness LIBRARY + its unit tests into
tests/extreme_harness/ so the default 'just test'/CI suite gates them
(2082 -> 2104 passed). tests-extensive/ now holds only the genuinely
harness-requiring scaffolding (compose, fixtures, run scripts) + the slow
marked tests, which consume the library.

- fault_proxy.py + its Dockerfile colocate in tests/extreme_harness/; the
  fault-proxy compose build context repoints there (verified: image builds).
- test_extreme_functional.py + the two cinefile run scripts import the library
  from tests.extreme_harness.
- Drop the now-obsolete test-extensive-unit recipe (no unmarked tests remain).
