+++ tests-extensive/extreme/scripts/cinefile_proxy_swap_loop.py
@@ -34,9 +34,10 @@ import urllib.parse
 import urllib.request
 from pathlib import Path
 
-# Local sibling import — share the PROPFIND helper with the other
-# runners so all three pick storages the same way.
-sys.path.insert(0, str(Path(__file__).resolve().parent))
+# Shared PROPFIND helper from the harness library so all the runners
+# pick storages the same way.
+_harness_dir = Path(__file__).resolve().parents[3] / "tests" / "extreme_harness"
+sys.path.insert(0, str(_harness_dir))
 from _storage_discovery import discover_cinefile_storages  # noqa: E402
 
 KODI_URL = os.environ.get("KODI_URL", "http://localhost:8082").rstrip("/")
